requirements.txt
absl-py==2.1.0
addict==2.4.0
albumentations==1.3.1
aliyun-python-sdk-core==2.15.2
aliyun-python-sdk-kms==2.16.5
antlr4-python3-runtime==4.9.3
attrdict==2.0.1
av==10.0.0
basicsr==1.4.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
coverage==7.2.7
crcmod==1.7
cryptography==43.0.1
cupy-cuda100==9.6.0
cycler==0.11.0
Cython==3.0.11
easydict==1.13
einops==0.6.1
exceptiongroup==1.2.2
facexlib==0.3.0
fastrlock==0.8.2
filelock==3.12.2
filterpy==1.4.5
flatbuffers==24.3.25
fonttools==4.38.0
fsspec==2023.1.0
ftfy==6.1.1
future==1.0.0
google-auth==2.34.0
google-auth-oauthlib==0.4.6
GPUtil==1.4.0
grpcio==1.36.1
h5py==3.8.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.8
imagecodecs==2021.11.20
imageio==2.31.2
imgaug==0.4.0
importlib-metadata==6.7.0
iniconfig==2.0.0
jmespath==0.10.0
joblib==1.3.2
kiwisolver==1.4.5
llvmlite==0.39.1
lmdb==1.5.1
loguru==0.7.2
lpips==0.1.4
mahotas==1.4.18
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.5.3
mdurl==0.1.2
# Editable install with no version control (mmcv-full==1.7.2)
-e /home/zwxionggroup/shoujt/envs/mmcv-1.x
mmedit==0.16.1
model-index==0.1.11
monai==1.1.0
mpmath==1.3.0
networkx==2.6.3
ninja==1.11.1.1
numba==0.56.4
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1649806299270/work
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.12.0
onnxoptimizer==0.3.13
onnxruntime==1.14.1
openai-clip==1.0.1
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.10
ordered-set==4.1.0
oss2==2.17.0
packaging==24.0
pandas==1.3.5
Pillow==9.5.0
platformdirs==4.0.0
pluggy==1.2.0
protobuf==3.19.6
psutil==6.0.0
ptflops==0.7.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
Pygments==2.17.2
pyiqa==0.1.10
pyparsing==3.1.4
pytest==7.4.4
pytest-runner==6.0.1
python-dateutil==2.9.0.post0
pytorch-msssim==1.0.0
PyTurboJPEG==1.7.6
pytz==2023.4
PyWavelets==1.3.0
PyYAML==6.0.1
qudida==0.0.4
regex==2024.4.16
requests==2.28.2
requests-oauthlib==2.0.0
rich==13.8.0
rsa==4.9
safetensors==0.4.4
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
shapely==2.0.6
six==1.16.0
sympy==1.10.1
tabulate==0.9.0
tb-nightly @ file:///home/zwxionggroup/shoujt/tb_nightly-2.12.0a20230113-py3-none-any.whl
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
termcolor==2.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.9.12
tomli==2.0.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.2
typing_extensions==4.7.1
urllib3==1.26.20
warmup-scheduler @ file:///home/zwxionggroup/shoujt/envs/warmup_scheduler-0.3.tar.gz
waterz==0.9.4
wcwidth==0.2.13
Werkzeug==2.2.3
yacs==0.1.8
yapf==0.40.2
zipp==3.15.0
